Addressed comments and added functional test
