--- lom/src/plugins/sonic/plugin_integration_tests/detection/link/linkcrc/LINK_CRC_COUNTERS3.txt
@@ -1,6 +1,0 @@
-{
-  "SAI_PORT_STAT_IF_IN_ERRORS": 850,
-  "SAI_PORT_STAT_IF_IN_UCAST_PKTS": 333,
-  "SAI_PORT_STAT_IF_OUT_UCAST_PKTS": 3100000000,
-  "SAI_PORT_STAT_IF_OUT_ERRORS": 3
-}
