+++ lom/src/plugins/sonic/plugin_integration_tests/linkcrc/LINK_CRC_COUNTERS1.txt
@@ -1,0 +1,6 @@
+{
+  "SAI_PORT_STAT_IF_IN_ERRORS": 100,
+  "SAI_PORT_STAT_IF_IN_UCAST_PKTS": 101,
+  "SAI_PORT_STAT_IF_OUT_UCAST_PKTS": 1100,
+  "SAI_PORT_STAT_IF_OUT_ERRORS": 1
+}
