+++ lom/src/plugins/sonic/plugin_integration_tests/linkcrc/LINK_CRC_COUNTERS4.txt
@@ -1,0 +1,6 @@
+{
+  "SAI_PORT_STAT_IF_IN_ERRORS": 1220,
+  "SAI_PORT_STAT_IF_IN_UCAST_PKTS": 444,
+  "SAI_PORT_STAT_IF_OUT_UCAST_PKTS": 4100000000,
+  "SAI_PORT_STAT_IF_OUT_ERRORS": 4
+}
