+++ lom/src/plugins/sonic/plugin_integration_tests/linkcrc/LINK_CRC_COUNTERS5.txt
@@ -1,0 +1,6 @@
+{
+  "SAI_PORT_STAT_IF_IN_ERRORS": 1650,
+  "SAI_PORT_STAT_IF_IN_UCAST_PKTS": 555,
+  "SAI_PORT_STAT_IF_OUT_UCAST_PKTS": 4100004000,
+  "SAI_PORT_STAT_IF_OUT_ERRORS": 5
+}
